backend/requirements.txt
flask
numpy
polars
pillow
matplotlib
pyyaml
